refactor(proxy): consolidate encrypt_config tests and add real cast test
Move the test-only config.rs module into manager.rs alongside the code
under test, extract a shared canonical_to_map helper used by both
load_encrypt_config and the tests, and add coverage for the `real`
cast_as alias that maps to ColumnType::Float.

Addresses PR #386 review feedback.
