+++ packages/cipherstash-proxy/src/proxy/encrypt_config/manager.rs
@@ -213,14 +213,7 @@ pub async fn load_encrypt_config(config: &DatabaseConfig) -> Result<EncryptConfi
 
             let json_value: Value = row.get("data");
             let canonical: CanonicalEncryptionConfig = serde_json::from_value(json_value)?;
-            let config_map = canonical
-                .into_config_map()
-                .map_err(crate::error::ConfigError::from)?;
-            let config_map = config_map
-                .into_iter()
-                .map(|(id, col)| (eql::Identifier::new(id.table, id.column), col))
-                .collect();
-            let encrypt_config = EncryptConfig::new_from_config(config_map);
+            let encrypt_config = EncryptConfig::new_from_config(canonical_to_map(canonical)?);
 
             Ok(encrypt_config)
         }
@@ -235,4 +228,451 @@ pub async fn load_encrypt_config(config: &DatabaseConfig) -> Result<EncryptConfi
 fn configuration_table_not_found(e: &tokio_postgres::Error) -> bool {
     let msg = e.to_string();
     msg.contains("eql_v2_configuration") && msg.contains("does not exist")
+}
+
+fn canonical_to_map(canonical: CanonicalEncryptionConfig) -> Result<EncryptConfigMap, ConfigError> {
+    Ok(canonical
+        .into_config_map()?
+        .into_iter()
+        .map(|(id, col)| (eql::Identifier::new(id.table, id.column), col))
+        .collect())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use cipherstash_client::eql::Identifier;
+    use cipherstash_config::column::{ArrayIndexMode, IndexType, TokenFilter, Tokenizer};
+    use cipherstash_config::ColumnType;
+    use serde_json::json;
+
+    fn parse(json: serde_json::Value) -> EncryptConfigMap {
+        let config: CanonicalEncryptionConfig = serde_json::from_value(json).unwrap();
+        canonical_to_map(config).unwrap()
+    }
+
+    #[test]
+    fn column_with_empty_options_gets_defaults() {
+        let json = json!({
+            "v": 1,
+            "tables": { "users": { "email": {} } }
+        });
+
+        let encrypt_config = parse(json);
+        let column = encrypt_config
+            .get(&Identifier::new("users", "email"))
+            .unwrap();
+
+        assert_eq!(column.cast_type, ColumnType::Text);
+        assert!(column.indexes.is_empty());
+    }
+
+    #[test]
+    fn can_parse_column_with_cast_as() {
+        let json = json!({
+            "v": 1,
+            "tables": {
+                "users": { "favourite_int": { "cast_as": "int" } }
+            }
+        });
+
+        let encrypt_config = parse(json);
+        let column = encrypt_config
+            .get(&Identifier::new("users", "favourite_int"))
+            .unwrap();
+
+        assert_eq!(column.cast_type, ColumnType::Int);
+        assert_eq!(column.name, "favourite_int");
+        assert!(column.indexes.is_empty());
+    }
+
+    #[test]
+    fn cast_as_real_maps_to_float() {
+        let json = json!({
+            "v": 1,
+            "tables": {
+                "users": { "rating": { "cast_as": "real" } }
+            }
+        });
+
+        let encrypt_config = parse(json);
+        let column = encrypt_config
+            .get(&Identifier::new("users", "rating"))
+            .unwrap();
+
+        assert_eq!(column.cast_type, ColumnType::Float);
+    }
+
+    #[test]
+    fn cast_as_double_maps_to_float() {
+        let json = json!({
+            "v": 1,
+            "tables": {
+                "users": { "rating": { "cast_as": "double" } }
+            }
+        });
+
+        let encrypt_config = parse(json);
+        let column = encrypt_config
+            .get(&Identifier::new("users", "rating"))
+            .unwrap();
+
+        assert_eq!(column.cast_type, ColumnType::Float);
+    }
+
+    #[test]
+    fn can_parse_empty_indexes() {
+        let json = json!({
+            "v": 1,
+            "tables": {
+                "users": { "email": { "indexes": {} } }
+            }
+        });
+
+        let encrypt_config = parse(json);
+        let column = encrypt_config
+            .get(&Identifier::new("users", "email"))
+            .unwrap();
+
+        assert!(column.indexes.is_empty());
+    }
+
+    #[test]
+    fn can_parse_ore_index() {
+        let json = json!({
+            "v": 1,
+            "tables": {
+                "users": { "email": { "indexes": { "ore": {} } } }
+            }
+        });
+
+        let encrypt_config = parse(json);
+        let column = encrypt_config
+            .get(&Identifier::new("users", "email"))
+            .unwrap();
+
+        assert_eq!(column.indexes[0].index_type, IndexType::Ore);
+    }
+
+    #[test]
+    fn can_parse_unique_index_with_defaults() {
+        let json = json!({
+            "v": 1,
+            "tables": {
+                "users": { "email": { "indexes": { "unique": {} } } }
+            }
+        });
+
+        let encrypt_config = parse(json);
+        let column = encrypt_config
+            .get(&Identifier::new("users", "email"))
+            .unwrap();
+
+        assert_eq!(
+            column.indexes[0].index_type,
+            IndexType::Unique {
+                token_filters: vec![]
+            }
+        );
+    }
+
+    #[test]
+    fn can_parse_unique_index_with_token_filter() {
+        let json = json!({
+            "v": 1,
+            "tables": {
+                "users": {
+                    "email": {
+                        "indexes": {
+                            "unique": {
+                                "token_filters": [{ "kind": "downcase" }]
+                            }
+                        }
+                    }
+                }
+            }
+        });
+
+        let encrypt_config = parse(json);
+        let column = encrypt_config
+            .get(&Identifier::new("users", "email"))
+            .unwrap();
+
+        assert_eq!(
+            column.indexes[0].index_type,
+            IndexType::Unique {
+                token_filters: vec![TokenFilter::Downcase]
+            }
+        );
+    }
+
+    #[test]
+    fn can_parse_match_index_with_defaults() {
+        let json = json!({
+            "v": 1,
+            "tables": {
+                "users": { "email": { "indexes": { "match": {} } } }
+            }
+        });
+
+        let encrypt_config = parse(json);
+        let column = encrypt_config
+            .get(&Identifier::new("users", "email"))
+            .unwrap();
+
+        assert_eq!(
+            column.indexes[0].index_type,
+            IndexType::Match {
+                tokenizer: Tokenizer::Standard,
+                token_filters: vec![],
+                k: 6,
+                m: 2048,
+                include_original: false,
+            }
+        );
+    }
+
+    #[test]
+    fn can_parse_match_index_with_all_opts_set() {
+        let json = json!({
+            "v": 1,
+            "tables": {
+                "users": {
+                    "email": {
+                        "indexes": {
+                            "match": {
+                                "tokenizer": { "kind": "ngram", "token_length": 3 },
+                                "token_filters": [{ "kind": "downcase" }],
+                                "k": 8,
+                                "m": 1024,
+                                "include_original": true
+                            }
+                        }
+                    }
+                }
+            }
+        });
+
+        let encrypt_config = parse(json);
+        let column = encrypt_config
+            .get(&Identifier::new("users", "email"))
+            .unwrap();
+
+        assert_eq!(
+            column.indexes[0].index_type,
+            IndexType::Match {
+                tokenizer: Tokenizer::Ngram { token_length: 3 },
+                token_filters: vec![TokenFilter::Downcase],
+                k: 8,
+                m: 1024,
+                include_original: true,
+            }
+        );
+    }
+
+    #[test]
+    fn can_parse_ste_vec_index() {
+        let json = json!({
+            "v": 1,
+            "tables": {
+                "users": {
+                    "event_data": {
+                        "cast_as": "jsonb",
+                        "indexes": { "ste_vec": { "prefix": "event-data" } }
+                    }
+                }
+            }
+        });
+
+        let encrypt_config = parse(json);
+        let column = encrypt_config
+            .get(&Identifier::new("users", "event_data"))
+            .unwrap();
+
+        assert_eq!(
+            column.indexes[0].index_type,
+            IndexType::SteVec {
+                prefix: "event-data".into(),
+                term_filters: vec![],
+                array_index_mode: ArrayIndexMode::ALL,
+            },
+        );
+    }
+
+    #[test]
+    fn config_map_preserves_table_and_column_names() {
+        let json = json!({
+            "v": 1,
+            "tables": {
+                "my_schema.users": {
+                    "email_address": {
+                        "cast_as": "text",
+                        "indexes": { "unique": {} }
+                    }
+                }
+            }
+        });
+
+        let config = parse(json);
+        let column = config
+            .get(&Identifier::new("my_schema.users", "email_address"))
+            .unwrap();
+        assert_eq!(column.name, "email_address");
+        assert_eq!(column.cast_type, ColumnType::Text);
+    }
+
+    #[test]
+    fn config_map_handles_multiple_tables() {
+        let json = json!({
+            "v": 1,
+            "tables": {
+                "users": { "email": { "cast_as": "text" } },
+                "orders": { "total": { "cast_as": "int" } }
+            }
+        });
+
+        let config = parse(json);
+
+        assert_eq!(config.len(), 2);
+        assert_eq!(
+            config
+                .get(&Identifier::new("users", "email"))
+                .unwrap()
+                .cast_type,
+            ColumnType::Text
+        );
+        assert_eq!(
+            config
+                .get(&Identifier::new("orders", "total"))
+                .unwrap()
+                .cast_type,
+            ColumnType::Int
+        );
+    }
+
+    #[test]
+    fn invalid_config_returns_error() {
+        let json = json!({
+            "v": 1,
+            "tables": {
+                "users": {
+                    "email": {
+                        "cast_as": "text",
+                        "indexes": { "ste_vec": { "prefix": "test" } }
+                    }
+                }
+            }
+        });
+
+        let config: CanonicalEncryptionConfig = serde_json::from_value(json).unwrap();
+        assert!(canonical_to_map(config).is_err());
+    }
+
+    #[test]
+    fn real_eql_config_produces_correct_encrypt_config() {
+        let json = json!({
+            "v": 1,
+            "tables": {
+                "encrypted": {
+                    "encrypted_text": {
+                        "cast_as": "text",
+                        "indexes": { "unique": {}, "match": {}, "ore": {} }
+                    },
+                    "encrypted_bool": {
+                        "cast_as": "boolean",
+                        "indexes": { "unique": {}, "ore": {} }
+                    },
+                    "encrypted_int2": {
+                        "cast_as": "small_int",
+                        "indexes": { "unique": {}, "ore": {} }
+                    },
+                    "encrypted_int4": {
+                        "cast_as": "int",
+                        "indexes": { "unique": {}, "ore": {} }
+                    },
+                    "encrypted_int8": {
+                        "cast_as": "big_int",
+                        "indexes": { "unique": {}, "ore": {} }
+                    },
+                    "encrypted_float8": {
+                        "cast_as": "double",
+                        "indexes": { "unique": {}, "ore": {} }
+                    },
+                    "encrypted_date": {
+                        "cast_as": "date",
+                        "indexes": { "unique": {}, "ore": {} }
+                    },
+                    "encrypted_jsonb": {
+                        "cast_as": "jsonb",
+                        "indexes": {
+                            "ste_vec": { "prefix": "encrypted/encrypted_jsonb" }
+                        }
+                    },
+                    "encrypted_jsonb_filtered": {
+                        "cast_as": "jsonb",
+                        "indexes": {
+                            "ste_vec": {
+                                "prefix": "encrypted/encrypted_jsonb_filtered",
+                                "term_filters": [{ "kind": "downcase" }]
+                            }
+                        }
+                    }
+                }
+            }
+        });
+
+        let config = parse(json);
+
+        assert_eq!(config.len(), 9);
+
+        assert_eq!(
+            config
+                .get(&Identifier::new("encrypted", "encrypted_float8"))
+                .unwrap()
+                .cast_type,
+            ColumnType::Float
+        );
+        assert_eq!(
+            config
+                .get(&Identifier::new("encrypted", "encrypted_jsonb"))
+                .unwrap()
+                .cast_type,
+            ColumnType::Json
+        );
+        assert_eq!(
+            config
+                .get(&Identifier::new("encrypted", "encrypted_text"))
+                .unwrap()
+                .indexes
+                .len(),
+            3
+        );
+        assert_eq!(
+            config
+                .get(&Identifier::new("encrypted", "encrypted_bool"))
+                .unwrap()
+                .indexes
+                .len(),
+            2
+        );
+        assert_eq!(
+            config
+                .get(&Identifier::new("encrypted", "encrypted_jsonb_filtered"))
+                .unwrap()
+                .indexes
+                .len(),
+            1
+        );
+    }
+
+    #[test]
+    fn malformed_json_returns_parse_error() {
+        let json = json!({
+            "v": 1,
+            "tables": "not a map"
+        });
+
+        let result = serde_json::from_value::<CanonicalEncryptionConfig>(json);
+        assert!(result.is_err());
+    }
 }
