+++ packages/cipherstash-proxy/src/proxy/encrypt_config/mod.rs
@@ -1,4 +1,3 @@
-mod config;
 mod manager;
 
 pub use manager::{EncryptConfig, EncryptConfigManager};
